Realizando ajustes de melhoria

+++ Api/SGH/Data/Repositorio/Contratos/IAulaRepositorio.cs
@@ -12,8 +12,8 @@ namespace SGH.Data.Repositorio.Contratos
         Task<bool> Contem(Expression<Func<Aula, bool>> expressao);
         Task<List<Aula>> Listar(Expression<Func<Aula, bool>> expressao);
         Task<List<Aula>> ListarComDisciplinas(Expression<Func<Aula, bool>> expressao);              
-        Task<bool> VerificarDisponibilidadeCargo(int codigoCargo, string diaSemana, string hora);
-        Task<bool> VerificarDisponibilidadeProfessor(int codigoProfessor, string diaSemana, string hora);
+        Task<bool> VerificarDisponibilidadeCargo(long codigoCargo, string diaSemana, string hora);
+        Task<bool> VerificarDisponibilidadeProfessor(long codigoProfessor, string diaSemana, string hora);
         Task<bool> Remover(Expression<Func<Aula, bool>> expressao);
     }
 }
