+++ Api/SGH/Data/Repositorio/Contratos/ICurriculoRepositorio.cs
@@ -9,13 +9,13 @@ namespace SGH.Data.Repositorio.Contratos
     public interface ICurriculoRepositorio : IRepositorioPaginacao<Curriculo>
     {
         Task<int> RetornarQuantidadeDisciplinaCurriculo(long codigoCurriculo);
-        Task<bool> Remover(int codigoCurriculo);
+        Task<bool> Remover(long codigoCurriculo);
         Task<Curriculo> Atualizar(Curriculo entidade);
         Task<Curriculo> Criar(Curriculo entidade);
         Task<bool> Contem(Expression<Func<Curriculo, bool>> expressao);
         Task<List<Curriculo>> ListarTodos();
         Task<Curriculo> Consultar(Expression<Func<Curriculo, bool>> expressao);
-        Task<IList<int>> ListarCodigos(Expression<Func<Curriculo, bool>> expressao);
-        Task<Curso> ConsultarCurso(int codigoCurriculo);
+        Task<IList<long>> ListarCodigos(Expression<Func<Curriculo, bool>> expressao);
+        Task<Curso> ConsultarCurso(long codigoCurriculo);
     }
 }
