+++ Api/SGH/Dominio/Contratos/IUsuarioComando.cs
@@ -2,11 +2,11 @@
 {
     public interface IUsuarioComando
     {
-        int? Codigo { get; set; }
+        long? Codigo { get; set; }
         string Email { get; set; }
         string Login { get; set; }
         string Nome { get; set; }
-        int PerfilCodigo { get; set; }
-        int? CursoCodigo { get; set; }
+        long PerfilCodigo { get; set; }
+        long? CursoCodigo { get; set; }
     }
 }
