+++ Api/SGH/Dominio/Implementacao/DIsciplinasTipo/Comandos/Remover/RemoverDisciplinaTipoComando.cs
@@ -1,11 +1,10 @@
 ﻿using MediatR;
 using SGH.Dominio.Core;
-using SGH.Dominio.Core.Model;
 
 namespace SGH.Dominio.Services.Implementacao.DIsciplinasTipoServico.Comandos.Remover
 {
     public class RemoverDisciplinaTipoComando: IRequest<Resposta<bool>>
     {
-        public int Codigo { get; set; }
+        public long Codigo { get; set; }
     }
 }
