+++ Api/SGH/Dominio/Implementacao/Usuarios/Comandos/Atualizar/AtualizarUsuarioComando.cs
@@ -7,15 +7,15 @@ namespace SGH.Dominio.Services.Implementacao.Usuarios.Comandos.Atualizar
 {
     public class AtualizarUsuarioComando : IUsuarioComando, IRequest<Resposta<Usuario>>
     {
-        public int? Codigo { get; set; }
+        public long? Codigo { get; set; }
         public string Nome { get; set; }
         public string Telefone { get; set; }
         public string Senha { get; set; }
-        public int PerfilCodigo { get; set; }
+        public long PerfilCodigo { get; set; }
         public string Foto { get; set; }
         public bool Ativo { get; set; }
         public string Email { get; set; }
         public string Login { get; set; }
-        public int? CursoCodigo { get; set; }
+        public long? CursoCodigo { get; set; }
     }
 }
