+++ Api/SGH/Dominio/ViewModel/CargoDisciplinaViewModel.cs
@@ -2,13 +2,13 @@
 {
     public class CargoDisciplinaViewModel
     {
-        public int? Codigo { get; set; }
-        public int? CodigoCargo { get; set; }
-        public int? CodigoCurriculoDisciplina { get; set; }
-        public int? CodigoCurriculo { get; set; }
+        public long? Codigo { get; set; }
+        public long? CodigoCargo { get; set; }
+        public long? CodigoCurriculoDisciplina { get; set; }
+        public long? CodigoCurriculo { get; set; }
         public string CursoDescricao { get; set; }
         public string Descricao { get; set; }
-        public int CodigoTurno { get; set; }
+        public long CodigoTurno { get; set; }
         public string TurnoDescricao { get; set; }
     }
 }
