+++ Api/SGH/SGH.Relatorios/DataSets/QuadroHorarioData.cs
@@ -2,7 +2,7 @@
 {
     public class QuadroHorarioData
     {
-        public int Codigo { get; set; }
+        public long Codigo { get; set; }
         public string Periodo { get; set; }
         public string Avisos { get; set; }
     }
