+++ Api/SGH/Servicos/Model/CargoDisciplina.cs
@@ -4,9 +4,9 @@ namespace SGH.Dominio.Core.Model
 {
     public class CargoDisciplina : EntidadeBase
     {
-        public int CodigoCurriculoDisciplina { get; set; }
-        public int CodigoCargo { get; set; }
-        public int CodigoTurno { get; set; }
+        public long CodigoCurriculoDisciplina { get; set; }
+        public long CodigoCargo { get; set; }
+        public long CodigoTurno { get; set; }
         public string Descricao { get; set; }
         public virtual Turno Turno { get; set; }
         public virtual CurriculoDisciplina Disciplina { get; set; }
