+++ Api/SGH/Servicos/Model/Curriculo.cs
@@ -4,8 +4,8 @@ namespace SGH.Dominio.Core.Model
 {
     public class Curriculo: EntidadeBase
     {
-        public int CodigoCurso { get; set; }
-        public int Ano { get; set; }
+        public long CodigoCurso { get; set; }
+        public long Ano { get; set; }
 
         public virtual Curso Curso { get; set; }
         public virtual IEnumerable<CurriculoDisciplina> Disciplinas { get; set; }
