+++ Api/SGH/Servicos/Model/HorarioAula.cs
@@ -8,8 +8,8 @@ namespace SGH.Dominio.Core.Model
         public int Ano { get; set; }
         public ESemestre Semestre { get; set; }
         public EPeriodo Periodo { get; set; }
-        public int CodigoTurno { get; set; }
-        public int CodigoCurriculo { get; set; }
+        public long CodigoTurno { get; set; }
+        public long CodigoCurriculo { get; set; }
         public string Mensagem { get; set; }
         public virtual Turno Turno { get; set; }
         public virtual Curriculo Curriculo { get; set; }
