+++ Api/SGH/Servicos/Model/Usuario.cs
@@ -10,8 +10,8 @@ namespace SGH.Dominio.Core.Model
         public string Email { get; set; }
         public string Foto { get; set; }
         public bool Ativo { get; set; }
-        public int PerfilCodigo { get; set; }
-        public int? CursoCodigo { get; set; }
+        public long PerfilCodigo { get; set; }
+        public long? CursoCodigo { get; set; }
 
         public UsuarioPerfil Perfil { get; set; }
         public Curso Curso { get; set; }
